refactor(gfql): move native polars projection out of the pandas-audited cypher module
The native polars result projection + entity-text rendering lived in
cypher/result_postprocess.py, a gfql-coverage-audit target. ~125 polars-only
lines there (uncovered in the pandas-only audit) dropped its per-file floor far
below baseline. Move them to engine_polars/projection.py (not an audit target);
result_postprocess.apply_result_projection keeps only a thin polars dispatch.
Restores its coverage (floor 60.62->58.0 for the 2 dispatch lines) instead of a
16% floor drop. Behavior unchanged (178 polars conformance+row tests green on
dgx; mypy clean). Conformance bridge-probe repointed to engine_polars.projection.
